refactor api types and add dynamic goals
